requirements.txt
python-dotenv
together
llama_index
faiss-cpu
setuptools
langchain==0.0.295